Add workaround for Ubuntu 23.10
Ubuntu 23.10 is slowly choking off our access to unprivileged user namespaces.
In particular, they have made it so that `unconfined` isn't actually unconfined,
and is unable to create unprivileged user namespaces.  This is super frustrating,
because there's no way to launch a process using unpriv userns without needing
superuser intervention (e.g. to create an apparmor profile).  Furthermore, it's
annoying to even install one, as the path to our executable is unreliable, since
it's in a content-addressed location.

And so, I regret more and more every day that I don't just give up and reimplement
everything on top of podman.  Until then, just ask the user to turn off these
annoying protections: